Fixed CreateIndex (with unique)
[git-p4: depot-paths = "//depot/qt/main/": change = 36246]

+++ tests/qdb/src/parser.cpp
@@ -1365,7 +1365,7 @@ void Parser::matchCreateStatement()
 	    ++col;
 	}
 	yyProg->append( new MakeList(columns.count()) );
-	yyProg->append( new CreateIndex(tableId, (int) unique) );
+	yyProg->append( new CreateIndex(tableId, unique) );
 	break;
     case Tok_table:
 	yyTok = getToken();
